+++ 4k.c
@@ -66,21 +66,30 @@ G(
     })
 
 G(
-    1,
-    typedef struct [[nodiscard]] {
-      ssize_t tv_sec;  // seconds
-      ssize_t tv_nsec; // nanoseconds
-    } timespec;
+    1, // Non-standard, gets but a word instead of a line
+    S(0) bool getl(char *restrict string) {
+      while (true) {
+        ssize_t result;
+        asm volatile("syscall"
+                     : "=a"(result)
+                     : "0"(0), "D"(stdin), "S"(string), "d"(1)
+                     : "rcx", "r11", "memory");
 
-    [[nodiscard]] S(0) u64 get_time() {
-      timespec ts;
-      ssize_t ret; // Unused
-      asm volatile("syscall"
-                   : "=a"(ret)
-                   : "0"(228), "D"(1), "S"(&ts)
-                   : "rcx", "r11", "memory");
-      return G(2, ts.tv_nsec) +
-             G(2, G(3, ts.tv_sec) * G(3, 1000 * 1000 * 1000ULL));
+    // Assume stdin never closes on mini build
+#ifdef FULL
+        if (result < 1) {
+          exit_now();
+        }
+#endif
+
+        const char ch = *string;
+        if (ch <= ' ') {
+          *string = 0;
+          return ch != '\n';
+        }
+
+        string++;
+      }
     })
 
 G(
@@ -116,30 +125,21 @@ G(
     })
 
 G(
-    1, // Non-standard, gets but a word instead of a line
-    S(0) bool getl(char *restrict string) {
-      while (true) {
-        ssize_t result;
-        asm volatile("syscall"
-                     : "=a"(result)
-                     : "0"(0), "D"(stdin), "S"(string), "d"(1)
-                     : "rcx", "r11", "memory");
+    1,
+    typedef struct [[nodiscard]] {
+      ssize_t tv_sec;  // seconds
+      ssize_t tv_nsec; // nanoseconds
+    } timespec;
 
-    // Assume stdin never closes on mini build
-#ifdef FULL
-        if (result < 1) {
-          exit_now();
-        }
-#endif
-
-        const char ch = *string;
-        if (ch <= ' ') {
-          *string = 0;
-          return ch != '\n';
-        }
-
-        string++;
-      }
+    [[nodiscard]] S(0) u64 get_time() {
+      timespec ts;
+      ssize_t ret; // Unused
+      asm volatile("syscall"
+                   : "=a"(ret)
+                   : "0"(228), "D"(1), "S"(&ts)
+                   : "rcx", "r11", "memory");
+      return G(2, ts.tv_nsec) +
+             G(2, G(3, ts.tv_sec) * G(3, 1000 * 1000 * 1000ULL));
     })
 
 #ifdef FULL
@@ -309,8 +309,8 @@ typedef struct [[nodiscard]] {
   G(5, u64 pieces[7];)
   G(5, u64 colour[2];)
   G(6, bool castling[4];)
-  G(6, bool flipped;)
   G(6, u8 padding[11];)
+  G(6, bool flipped;)
 } Position;
 
 #ifdef ASSERTS
@@ -364,9 +364,9 @@ G(
 
 G(
     16, [[nodiscard]] S(1) u64 southeast(const u64 bb) {
+      return G(17, G(18, east)(G(18, south)(bb)));
       return G(17, shift(H(10, 2, -7), H(10, 2, ~0x101010101010101ull),
                          H(10, 2, bb)));
-      return G(17, G(18, east)(G(18, south)(bb)));
     })
 
 G(
@@ -417,21 +417,21 @@ G(
 
 G(
     32, [[nodiscard]] S(0) u64 king(const u64 bb) {
-      const u64 vertical = G(38, north(bb)) | G(38, south(bb));
-      const u64 vertical_inclusive = G(39, bb) | G(39, vertical);
-      return G(40, vertical) | G(40, east(vertical_inclusive)) |
-             G(40, west(vertical_inclusive));
+      const u64 vertical = G(33, north(bb)) | G(33, south(bb));
+      const u64 vertical_inclusive = G(34, bb) | G(34, vertical);
+      return G(35, vertical) | G(35, east(vertical_inclusive)) |
+             G(35, west(vertical_inclusive));
     })
 
 G(
     32, [[nodiscard]] S(0) u64 knight(const u64 bb) {
-      G(33, const u64 east_bb = east(bb);)
-      G(33, const u64 west_bb = west(bb);)
-      G(34, const u64 horizontal1 = G(35, west_bb) | G(35, east_bb);)
-      G(34,
-        const u64 horizontal2 = G(36, east(east_bb)) | G(36, west(west_bb));)
-      return G(37, horizontal2 << 8) | G(37, horizontal2 >> 8) |
-             G(37, horizontal1 << 16) | G(37, horizontal1 >> 16);
+      G(36, const u64 east_bb = east(bb);)
+      G(36, const u64 west_bb = west(bb);)
+      G(37, const u64 horizontal1 = G(38, west_bb) | G(38, east_bb);)
+      G(37,
+        const u64 horizontal2 = G(39, east(east_bb)) | G(39, west(west_bb));)
+      return G(40, horizontal2 << 8) | G(40, horizontal2 >> 8) |
+             G(40, horizontal1 << 16) | G(40, horizontal1 >> 16);
     })
 
 G(
@@ -439,11 +439,11 @@ G(
                                     H(41, 1, const u64 bb)) {
       assert(count(bb) == 1);
       return G(42, // West
-               ray(H(26, 3, blockers), H(26, 3, ~0x8080808080808080ull),
-                   H(26, 3, bb), H(26, 3, -1))) |
+               ray(H(26, 2, blockers), H(26, 2, ~0x8080808080808080ull),
+                   H(26, 2, bb), H(26, 2, -1))) |
              G(42, // East
-               ray(H(26, 2, blockers), H(26, 2, ~0x101010101010101ull),
-                   H(26, 2, bb), H(26, 2, 1))) |
+               ray(H(26, 3, blockers), H(26, 3, ~0x101010101010101ull),
+                   H(26, 3, bb), H(26, 3, 1))) |
              G(42, xattack(H(27, 2, blockers), H(27, 2, bb),
                            H(27, 2, bb ^ 0x101010101010101ULL << lsb(bb) % 8)));
     })
@@ -474,8 +474,8 @@ G(
 
 G(
     46,
-    S(1) void move_str(H(54, 1, const Move *restrict move),
-                       H(54, 1, const i32 flip), H(54, 1, char *restrict str)) {
+    S(1) void move_str(H(50, 1, const Move *restrict move),
+                       H(50, 1, const i32 flip), H(50, 1, char *restrict str)) {
       assert(move->from >= 0);
       assert(move->from < 64);
       assert(move->to >= 0);
@@ -485,31 +485,31 @@ G(
              move->promo == Bishop || move->promo == Rook ||
              move->promo == Queen);
 
-      G(55, str[5] = '\0';)
+      G(51, str[5] = '\0';)
 
-      G(55, str[4] = "\0\0nbrq"[move->promo];)
+      G(51, str[4] = "\0\0nbrq"[move->promo];)
       G(
-          55, // Hack to save bytes, technically UB but works on GCC 14.2
+          51, // Hack to save bytes, technically UB but works on GCC 14.2
           for (i32 i = 0; i < 2; i++) {
-            G(56, str[i * 2 + 1] = '1' + ((&move->from)[i] / 8 ^ 7 * flip);)
-            G(56, str[i * 2] = 'a' + (&move->from)[i] % 8;)
+            G(52, str[i * 2 + 1] = '1' + ((&move->from)[i] / 8 ^ 7 * flip);)
+            G(52, str[i * 2] = 'a' + (&move->from)[i] % 8;)
           })
     })
 
 G(
-    46, [[nodiscard]] S(1) bool move_equal(G(50, Move *const rhs),
-                                           G(50, Move *const lhs)) {
-      return G(51, *(u32 *)lhs) == G(51, *(u32 *)rhs);
+    46, [[nodiscard]] S(1) bool move_equal(G(53, Move *const rhs),
+                                           G(53, Move *const lhs)) {
+      return G(54, *(u32 *)lhs) == G(54, *(u32 *)rhs);
     })
 
 G(
     46, [[nodiscard]] S(0)
-            i32 piece_on(H(52, 1, const Position *const restrict pos),
-                         H(52, 1, const i32 sq)) {
+            i32 piece_on(H(55, 1, const Position *const restrict pos),
+                         H(55, 1, const i32 sq)) {
               assert(sq >= 0);
               assert(sq < 64);
               for (i32 i = Pawn; i <= King; ++i) {
-                if (G(53, pos->pieces[i]) & G(53, 1ull << sq)) {
+                if (G(56, pos->pieces[i]) & G(56, 1ull << sq)) {
                   return i;
                 }
               }
@@ -527,35 +527,35 @@ G(
       G(59, const u64 blockers = theirs | pos->colour[0];)
       return G(60, G(61, (G(62, southwest(pawns)) | G(62, southeast(pawns)))) &
                        G(61, bb)) ||
-             G(60, G(66, rook(H(41, 2, blockers), H(41, 2, bb))) &
-                       G(66, (pos->pieces[Rook] | pos->pieces[Queen])) &
-                       G(66, theirs)) ||
+             G(60, G(63, rook(H(41, 2, blockers), H(41, 2, bb))) &
+                       G(63, (pos->pieces[Rook] | pos->pieces[Queen])) &
+                       G(63, theirs)) ||
              G(60,
-               G(63, pos->pieces[King]) & G(63, theirs) & G(63, king(bb))) ||
+               G(64, pos->pieces[King]) & G(64, theirs) & G(64, king(bb))) ||
              G(60, G(65, bishop(H(43, 2, blockers), H(43, 2, bb))) &
                        G(65, theirs) &
                        G(65, (pos->pieces[Bishop] | pos->pieces[Queen]))) ||
              G(60,
-               G(64, pos->pieces[Knight]) & G(64, knight(bb)) & G(64, theirs));
+               G(66, pos->pieces[Knight]) & G(66, knight(bb)) & G(66, theirs));
     })
 
 G(
     57, S(0) void flip_pos(Position *const restrict pos) {
+      G(67, pos->flipped ^= 1;)
       G(67, pos->colour[0] ^= pos->colour[1]; pos->colour[1] ^= pos->colour[0];
         pos->colour[0] ^= pos->colour[1];)
+
       G(
           67, // Hack to flip the first 10 bitboards in Position.
               // Technically UB but works in GCC 14.2
           u64 *pos_ptr = (u64 *)pos;
           for (i32 i = 0; i < 10; i++) { pos_ptr[i] = flip_bb(pos_ptr[i]); })
-
-      G(67, pos->flipped ^= 1;)
       G(67, u32 *c = (u32 *)pos->castling;
         *c = G(68, (*c >> 16)) | G(68, (*c << 16));)
     })
 
 G(
-    57, [[nodiscard]] S(1) u64 get_mobility(H(69, 1, const Position *pos),
+    57, [[nodiscard]] S(0) u64 get_mobility(H(69, 1, const Position *pos),
                                             H(69, 1, const i32 piece),
                                             H(69, 1, const i32 sq)) {
       u64 moves = 0;
@@ -610,7 +610,7 @@ G(
                       .from = from,
                       .to = to,
                       .promo = None,
-                      .takes_piece = piece_on(H(52, 2, pos), H(52, 2, to))});)
+                      .takes_piece = piece_on(H(55, 2, pos), H(55, 2, to))});)
           }
         }
       }
@@ -630,11 +630,11 @@ G(
              move->promo == Bishop || move->promo == Rook ||
              move->promo == Queen);
       assert(move->takes_piece != King);
-      assert(move->takes_piece == piece_on(H(52, 3, pos), H(52, 3, move->to)));
+      assert(move->takes_piece == piece_on(H(55, 3, pos), H(55, 3, move->to)));
 
       G(81, const u64 from = 1ull << move->from;)
       G(81, const u64 to = 1ull << move->to;)
-      G(82, const i32 piece = piece_on(H(52, 4, pos), H(52, 4, move->from));
+      G(82, const i32 piece = piece_on(H(55, 4, pos), H(55, 4, move->from));
         assert(piece != None);)
       G(82, const u64 mask = G(83, from) | G(83, to);)
 
@@ -644,8 +644,8 @@ G(
             const u64 bb = move->to - move->from == 2   ? 0xa0
                            : move->from - move->to == 2 ? 0x9
                                                         : 0;
-            G(85, pos->colour[0] ^= bb;)
             G(85, pos->pieces[Rook] ^= bb;)
+            G(85, pos->colour[0] ^= bb;)
           })
 
       G(84, pos->colour[0] ^= mask;)
@@ -681,10 +681,10 @@ G(
 
       G(89, // Update castling permissions
         const u64 oppMask = mask >> 56;
-        G(92, pos->castling[2] &= !(oppMask & 0x90);)
-            G(92, pos->castling[0] &= !(mask & 0x90);)
-                G(92, pos->castling[1] &= !(mask & 0x11);)
-                    G(92, pos->castling[3] &= !(oppMask & 0x11);))
+        G(92, pos->castling[3] &= !(oppMask & 0x11);)
+            G(92, pos->castling[1] &= !(mask & 0x11);)
+                G(92, pos->castling[0] &= !(mask & 0x90);)
+                    G(92, pos->castling[2] &= !(oppMask & 0x90);))
 
       if (find_in_check(pos)) {
         return false;
@@ -726,8 +726,8 @@ G(
         assert(from < 64);
         assert(to >= 0);
         assert(to < 64);
-        assert(piece_on(H(52, 5, pos), H(52, 5, from)) == Pawn);
-        const u8 takes = piece_on(H(52, 6, pos), H(52, 6, to));
+        assert(piece_on(H(55, 5, pos), H(55, 5, from)) == Pawn);
+        const u8 takes = piece_on(H(55, 6, pos), H(55, 6, to));
         if (to > 55) {
           for (u8 piece = Queen; piece >= Knight; piece--) {
             *movelist++ = ((Move){
@@ -766,19 +766,11 @@ enum { max_moves = 218 };
       })
   G(98, // PAWN DOUBLE MOVES
     movelist = generate_pawn_moves(
-        H(93, 4, -8),
-        H(93, 4,
-          north(G(105, G(106, pos->colour[0]) & G(106, pos->pieces[Pawn]))) &
-              G(105, ~all) &
-              G(105, (only_captures ? 0xFF00000000000000ull : ~0ull))),
-        H(93, 4, movelist), H(93, 4, pos));)
-  G(98, // PAWN EAST CAPTURES
-    movelist = generate_pawn_moves(
-        H(93, 3, -9),
+        H(93, 3, -8),
         H(93, 3,
-          G(102,
-            northeast(G(103, pos->colour[0]) & G(103, pos->pieces[Pawn]))) &
-              G(102, (G(104, pos->colour[1]) | G(104, pos->ep)))),
+          north(G(102, G(103, pos->colour[0]) & G(103, pos->pieces[Pawn]))) &
+              G(102, ~all) &
+              G(102, (only_captures ? 0xFF00000000000000ull : ~0ull))),
         H(93, 3, movelist), H(93, 3, pos));)
   G(98, // PAWN WEST CAPTURES
     movelist = generate_pawn_moves(
@@ -788,6 +780,14 @@ enum { max_moves = 218 };
             northwest(G(108, pos->colour[0]) & G(108, pos->pieces[Pawn]))) &
               G(107, (G(109, pos->colour[1]) | G(109, pos->ep)))),
         H(93, 5, movelist), H(93, 5, pos));)
+  G(98, // PAWN EAST CAPTURES
+    movelist = generate_pawn_moves(
+        H(93, 4, -9),
+        H(93, 4,
+          G(104,
+            northeast(G(105, pos->colour[0]) & G(105, pos->pieces[Pawn]))) &
+              G(104, (G(106, pos->colour[1]) | G(106, pos->ep)))),
+        H(93, 4, movelist), H(93, 4, pos));)
   G(
       98, // LONG CASTLE
       if (G(110, !only_captures) && G(110, !(G(111, all) & G(111, 0xEull))) &&
@@ -1063,15 +1063,16 @@ S(0) i32 eval(Position *const restrict pos) {
           pawns[i] = G(127, pos->colour[i]) & G(127, pos->pieces[Pawn]);
         } const u64 attacked_by_pawns = G(128, southwest(pawns[1])) |
                                         G(128, southeast(pawns[1]));
-        G(129, // PROTECTED PAWNS
-          score -= G(133, eval_params.protected_pawn) *
-                   G(133, count(G(134, pawns[1]) & G(134, attacked_by_pawns)));)
-            G(129, const u64 no_passers =
-                       G(130, pawns[1]) | G(130, attacked_by_pawns);)
-                G(129, // PHALANX PAWNS
+        G(129,
+          const u64 no_passers = G(132, pawns[1]) | G(132, attacked_by_pawns);)
+            G(129, // PHALANX PAWNS
+              score -=
+              G(133, eval_params.phalanx_pawn) *
+              G(133, count(G(134, pawns[1]) & G(134, west(pawns[1]))));)
+                G(129, // PROTECTED PAWNS
                   score -=
-                  G(131, eval_params.phalanx_pawn) *
-                  G(131, count(G(132, pawns[1]) & G(132, west(pawns[1]))));))
+                  G(130, eval_params.protected_pawn) *
+                  G(130, count(G(131, pawns[1]) & G(131, attacked_by_pawns)));))
     G(
         125, // BISHOP PAIR
         if (count(G(135, pos->pieces[Bishop]) & G(135, pos->colour[0])) > 1) {
@@ -1167,14 +1168,14 @@ S(0) i32 eval(Position *const restrict pos) {
 
                 G(166, // MOBILITY
                   score +=
-                  G(169, count(G(170, ~pos->colour[0]) & G(170, mobility) &
-                               G(170, ~attacked_by_pawns))) *
-                  G(169, eval_params.mobilities[p - 2]);)
+                  G(167, count(G(168, ~pos->colour[0]) & G(168, mobility) &
+                               G(168, ~attacked_by_pawns))) *
+                  G(167, eval_params.mobilities[p - 2]);)
 
                     G(166, // KING ATTACKS
                       score +=
-                      G(167, count(G(168, mobility) & G(168, opp_king_zone))) *
-                      G(167, eval_params.king_attacks[p - 2]);))
+                      G(169, count(G(170, mobility) & G(170, opp_king_zone))) *
+                      G(169, eval_params.king_attacks[p - 2]);))
             })
       }
     }
@@ -1211,18 +1212,18 @@ enum { corrhist_size = 16384 };
 
 typedef struct [[nodiscard]] {
   G(119, Move best_move;)
-  G(119, Move killer;)
+  G(119, i32 static_eval;)
   G(119, i32 num_moves;)
   G(119, u64 position_hash;)
-  G(119, i32 static_eval;)
+  G(119, Move killer;)
 } SearchStack;
 
 typedef struct [[nodiscard]] __attribute__((packed)) {
-  G(174, u16 partial_hash;)
   G(174, i16 score;)
-  G(174, i8 depth;)
   G(174, u8 flag;)
+  G(174, u16 partial_hash;)
   G(174, Move move;)
+  G(174, i8 depth;)
 } TTEntry;
 _Static_assert(sizeof(TTEntry) == 10);
 
@@ -1316,11 +1317,15 @@ get_hash(const Position *const pos) {
   u64 hash = 0;
   for (i32 c = 0; c < 2; c++) {
     for (i32 p = Pawn; p <= Queen; p++) {
-      hash = G(998, G(999, hash) * G(999, 9)) +
-             G(998, count(G(997, pos->pieces[p]) & G(997, pos->colour[c])));
+      hash = G(178, G(179, hash) * G(179, 9)) +
+             G(178, count(G(180, pos->pieces[p]) & G(180, pos->colour[c])));
     }
   }
   return hash;
+}
+
+[[nodiscard]] S(1) u64 get_pawn_hash(const Position *const pos) {
+  return (G(181, pos->pieces[Pawn]) * G(181, 0x9E3779B97F4A7C15ULL)) >> 50;
 }
 
 S(1)
@@ -1328,10 +1333,10 @@ i32 search(
 #ifdef FULL
     u64 *nodes,
 #endif
-    H(178, 1, H(179, 1, Position *const pos), H(179, 1, const i32 beta),
-      H(179, 1, i32 depth), H(179, 1, ThreadData *data)),
-    H(178, 1, H(180, 1, const i32 ply), H(180, 1, const bool do_null),
-      H(180, 1, i32 alpha))) {
+    H(182, 1, H(183, 1, Position *const pos), H(183, 1, const i32 beta),
+      H(183, 1, i32 depth), H(183, 1, ThreadData *data)),
+    H(182, 1, H(184, 1, const i32 ply), H(184, 1, const bool do_null),
+      H(184, 1, i32 alpha))) {
   assert(alpha < beta);
   assert(ply >= 0);
 
@@ -1345,23 +1350,23 @@ i32 search(
   // FULL REPETITION DETECTION
   const u64 tt_hash = get_hash(pos);
   bool in_qsearch = depth <= 0;
-  for (i32 i = G(181, ply); G(182, i >= 0) && G(182, do_null); i -= 2) {
-    if (G(183, tt_hash) == G(183, stack[i].position_hash)) {
+  for (i32 i = G(185, ply); G(186, i >= 0) && G(186, do_null); i -= 2) {
+    if (G(187, tt_hash) == G(187, stack[i].position_hash)) {
       return 0;
     }
   }
 
   // TT PROBING
-  G(184, TTEntry *tt_entry = &tt[tt_hash % tt_length];)
-  G(184, const u16 tt_hash_partial = tt_hash / tt_length;)
-  G(184, stack[ply].best_move = (Move){0};)
-  if (G(185, tt_entry->partial_hash) == G(185, tt_hash_partial)) {
+  G(188, TTEntry *tt_entry = &tt[tt_hash % tt_length];)
+  G(188, const u16 tt_hash_partial = tt_hash / tt_length;)
+  G(188, stack[ply].best_move = (Move){0};)
+  if (G(189, tt_entry->partial_hash) == G(189, tt_hash_partial)) {
     stack[ply].best_move = tt_entry->move;
 
     // TT PRUNING
-    if (G(186, tt_entry->depth >= depth) &&
-        G(186, G(187, tt_entry->flag) != G(187, tt_entry->score <= alpha)) &&
-        G(186, G(188, alpha) == G(188, beta - 1))) {
+    if (G(190, tt_entry->depth >= depth) &&
+        G(190, G(191, tt_entry->flag) != G(191, tt_entry->score <= alpha)) &&
+        G(190, G(192, alpha) == G(192, beta - 1))) {
       return tt_entry->score;
     }
   } else if (depth > 3) {
@@ -1371,119 +1376,122 @@ i32 search(
   }
 
   // STATIC EVAL WITH CORRECTION HISTORY
-  G(189, const i32 raw_eval = eval(pos); assert(raw_eval < mate);
-    assert(raw_eval > -mate);)
-
-  G(189, const u64 material_hash = get_material_hash(pos);
-    i32 *material_entry =
-        &data->corrhist[pos->flipped][material_hash % corrhist_size];)
-  i32 static_eval = G(190, raw_eval) + G(190, *material_entry / 256);
-  assert(static_eval < mate);
-  assert(static_eval > -mate);
+  G(193, i32 static_eval = eval(pos); assert(static_eval < mate);
+    assert(static_eval > -mate);)
+  G(193, const u64 corr_hashes[2] = {G(194, get_material_hash(pos)),
+                                     G(194, get_pawn_hash(pos))};)
+  G(193, i32 * corr_entries[2];)
+  for (i32 i = 0; i < 2; i++) {
+    corr_entries[i] =
+        &data->corrhist[pos->flipped][corr_hashes[i] % corrhist_size];
+    static_eval += *corr_entries[i] / 256;
+    assert(static_eval < mate);
+    assert(static_eval > -mate);
+  }
 
   stack[ply].static_eval = static_eval;
   const bool improving = ply > 1 && static_eval > stack[ply - 2].static_eval;
-  if (G(191, G(193, tt_entry->partial_hash) == G(193, tt_hash_partial)) &&
-      G(191, G(192, tt_entry->flag) != G(192, static_eval) > tt_entry->score)) {
+  if (G(195, G(196, tt_entry->partial_hash) == G(196, tt_hash_partial)) &&
+      G(195, G(197, tt_entry->flag) != G(197, static_eval) > tt_entry->score)) {
     static_eval = tt_entry->score;
   }
 
   // QUIESCENCE
-  if (G(194, static_eval > alpha) && G(194, in_qsearch)) {
+  if (G(198, static_eval > alpha) && G(198, in_qsearch)) {
     if (static_eval >= beta) {
       return static_eval;
     }
     alpha = static_eval;
   }
 
-  if (G(195, !in_check) && G(195, G(196, alpha) == G(196, beta - 1))) {
-    if (G(197, depth < 9) && G(197, !in_qsearch)) {
+  if (G(199, !in_check) && G(199, G(200, alpha) == G(200, beta - 1))) {
+    if (G(201, depth < 9) && G(201, !in_qsearch)) {
 
-      G(198, {
+      G(202, {
         // REVERSE FUTILITY PRUNING
-        if (static_eval - G(199, 57) * G(199, (depth - improving)) >= beta) {
+        if (static_eval - G(203, 57) * G(203, (depth - improving)) >= beta) {
           return static_eval;
         }
       })
 
-      G(198, // RAZORING
+      G(202, // RAZORING
         in_qsearch =
-            G(200, static_eval) + G(200, G(201, 115) * G(201, depth)) <= alpha;)
+            G(204, static_eval) + G(204, G(205, 115) * G(205, depth)) <= alpha;)
     }
 
     // NULL MOVE PRUNING
-    if (G(202, depth > 2) && G(202, do_null) && G(202, static_eval >= beta)) {
+    if (G(206, depth > 2) && G(206, do_null) && G(206, static_eval >= beta)) {
       Position npos = *pos;
-      G(203, flip_pos(&npos);)
-      G(203, npos.ep = 0;)
+      G(207, flip_pos(&npos);)
+      G(207, npos.ep = 0;)
       const i32 score = -search(
 #ifdef FULL
           nodes,
 #endif
-          H(178, 2, H(179, 2, &npos), H(179, 2, -alpha),
-            H(179, 2, depth - G(204, depth / 4) - G(204, 4)), H(179, 2, data)),
-          H(178, 2, H(180, 2, ply + 1), H(180, 2, false), H(180, 2, -beta)));
+          H(182, 2, H(183, 2, &npos), H(183, 2, -alpha),
+            H(183, 2, depth - G(208, depth / 4) - G(208, 4)), H(183, 2, data)),
+          H(182, 2, H(184, 2, ply + 1), H(184, 2, false), H(184, 2, -beta)));
       if (score >= beta) {
         return score;
       }
     }
   }
 
-  G(205, stack[G(206, ply) + G(206, 2)].position_hash = tt_hash;)
-  G(205, i32 quiets_evaluated = 0;)
-  G(205, i32 best_score = in_qsearch ? static_eval : -inf;)
-  G(205, Move moves[max_moves];
+  G(209, Move moves[max_moves];
     stack[ply].num_moves =
         movegen(H(95, 3, pos), H(95, 3, in_qsearch), H(95, 3, moves));)
-  G(205, i32 moves_evaluated = 0;)
-  G(205, u8 tt_flag = Upper;)
+  G(209, i32 best_score = in_qsearch ? static_eval : -inf;)
+  G(209, stack[G(210, ply) + G(210, 2)].position_hash = tt_hash;)
+  G(209, u8 tt_flag = Upper;)
+  G(209, i32 quiets_evaluated = 0;)
+  G(209, i32 moves_evaluated = 0;)
 
   for (i32 move_index = 0; move_index < stack[ply].num_moves; move_index++) {
     // MOVE ORDERING
-    G(207, i32 best_index = 0;)
-    G(207, i32 move_score = ~0x1010101LL;)
+    G(211, i32 move_score = ~0x1010101LL;)
+    G(211, i32 best_index = 0;)
     for (i32 order_index = move_index; order_index < stack[ply].num_moves;
          order_index++) {
       assert(moves[order_index].takes_piece ==
-             piece_on(H(52, 7, pos), H(52, 7, moves[order_index].to)));
+             piece_on(H(55, 7, pos), H(55, 7, moves[order_index].to)));
       const i32 order_move_score =
-          G(179, // KILLER MOVE
-            G(208, move_equal(G(209, &moves[order_index]),
-                              G(209, &stack[ply].killer))) *
-                G(208, 819)) +
-          G(179, // HISTORY HEURISTIC
+          G(183, // KILLER MOVE
+            G(212, move_equal(G(213, &moves[order_index]),
+                              G(213, &stack[ply].killer))) *
+                G(212, 819)) +
+          G(183, // HISTORY HEURISTIC
             move_history[pos->flipped][moves[order_index].takes_piece]
                         [moves[order_index].from][moves[order_index].to]) +
-          G(179, // PREVIOUS BEST MOVE FIRST
-            (move_equal(G(211, &stack[ply].best_move),
-                        G(211, &moves[order_index]))
+          G(183, // PREVIOUS BEST MOVE FIRST
+            (move_equal(G(214, &stack[ply].best_move),
+                        G(214, &moves[order_index]))
              << 30)) +
-          G(179, // MOST VALUABLE VICTIM
-            G(210, moves[order_index].takes_piece) * G(210, 622));
+          G(183, // MOST VALUABLE VICTIM
+            G(215, moves[order_index].takes_piece) * G(215, 622));
       if (order_move_score > move_score) {
-        G(212, best_index = order_index;)
-        G(212, move_score = order_move_score;)
+        G(216, best_index = order_index;)
+        G(216, move_score = order_move_score;)
       }
     }
 
-    swapmoves(G(213, &moves[best_index]), G(213, &moves[move_index]));
+    swapmoves(G(217, &moves[best_index]), G(217, &moves[move_index]));
 
     G(
-        214, // FORWARD FUTILITY PRUNING / DELTA PRUNING
-        if (G(217, depth < 8) &&
-            G(217,
-              G(218, static_eval) + G(218, G(219, 154) * G(219, depth)) +
-                      G(218,
+        218, // MOVE SCORE PRUNING
+        if (G(222, moves_evaluated) &&
+            G(222, move_score < G(223, -125) * G(223, depth))) { break; })
+
+    G(
+        218, // FORWARD FUTILITY PRUNING / DELTA PRUNING
+        if (G(219, depth < 8) &&
+            G(219,
+              G(220, static_eval) + G(220, G(221, 154) * G(221, depth)) +
+                      G(220,
                         initial_params.eg.material[moves[move_index].promo]) +
-                      G(218, initial_params.eg
+                      G(220, initial_params.eg
                                  .material[moves[move_index].takes_piece]) <
                   alpha) &&
-            G(217, moves_evaluated) && G(217, !in_check)) { break; })
-
-    G(
-        214, // MOVE SCORE PRUNING
-        if (G(215, moves_evaluated) &&
-            G(215, move_score < G(216, -125) * G(216, depth))) { break; })
+            G(219, moves_evaluated) && G(219, !in_check)) { break; })
 
     Position npos = *pos;
 #ifdef FULL
@@ -1498,10 +1506,10 @@ i32 search(
     moves_evaluated++;
 
     // LATE MOVE REDUCTION
-    i32 reduction = G(220, depth > 3) && G(220, move_score <= 0)
-                        ? G(221, !improving) + G(221, (move_score / -358)) +
-                              G(221, (G(222, alpha) == G(222, beta - 1))) +
-                              G(221, moves_evaluated / 9)
+    i32 reduction = G(224, depth > 3) && G(224, move_score <= 0)
+                        ? G(225, !improving) + G(225, (move_score / -358)) +
+                              G(225, (G(226, alpha) == G(226, beta - 1))) +
+                              G(225, moves_evaluated / 9)
                         : 0;
 
     i32 score;
@@ -1510,9 +1518,9 @@ i32 search(
 #ifdef FULL
           nodes,
 #endif
-          H(178, 3, H(179, 3, &npos), H(179, 3, -alpha),
-            H(179, 3, depth - G(223, 1) - G(223, reduction)), H(179, 3, data)),
-          H(178, 3, H(180, 3, ply + 1), H(180, 3, true), H(180, 3, low)));
+          H(182, 3, H(183, 3, &npos), H(183, 3, -alpha),
+            H(183, 3, depth - G(227, 1) - G(227, reduction)), H(183, 3, data)),
+          H(182, 3, H(184, 3, ply + 1), H(184, 3, true), H(184, 3, low)));
 
       // EARLY EXITS
       if (stop || (depth > 4 && get_time() - start_time > data->max_time)) {
@@ -1537,37 +1545,37 @@ i32 search(
       best_score = score;
 
       if (score > alpha) {
-        G(224, tt_flag = Exact;)
-        G(224, stack[ply].best_move = moves[move_index];)
-        G(224, alpha = score;)
+        G(228, tt_flag = Exact;)
+        G(228, stack[ply].best_move = moves[move_index];)
+        G(228, alpha = score;)
         if (score >= beta) {
           assert(stack[ply].best_move.takes_piece ==
-                 piece_on(H(52, 8, pos), H(52, 8, stack[ply].best_move.to)));
-          G(225, tt_flag = Lower;)
+                 piece_on(H(55, 8, pos), H(55, 8, stack[ply].best_move.to)));
+          G(229, tt_flag = Lower;)
           G(
-              225, if (stack[ply].best_move.takes_piece == None) {
+              229, if (stack[ply].best_move.takes_piece == None) {
                 stack[ply].killer = stack[ply].best_move;
               })
           G(
-              225, if (!in_qsearch) {
+              229, if (!in_qsearch) {
                 const i32 bonus = depth * depth;
-                G(226, i32 *const this_hist =
+                G(230, i32 *const this_hist =
                            &move_history[pos->flipped]
                                         [stack[ply].best_move.takes_piece]
                                         [stack[ply].best_move.from]
                                         [stack[ply].best_move.to];
 
                   *this_hist +=
-                  bonus - G(227, bonus) * G(227, *this_hist) / 1024;)
+                  bonus - G(231, bonus) * G(231, *this_hist) / 1024;)
                 G(
-                    226, for (i32 prev_index = 0; prev_index < move_index;
+                    230, for (i32 prev_index = 0; prev_index < move_index;
                               prev_index++) {
                       const Move prev = moves[prev_index];
                       i32 *const prev_hist =
                           &move_history[pos->flipped][prev.takes_piece]
                                        [prev.from][prev.to];
                       *prev_hist -=
-                          bonus + G(228, bonus) * G(228, *prev_hist) / 1024;
+                          bonus + G(232, bonus) * G(232, *prev_hist) / 1024;
                     })
               })
           break;
@@ -1580,17 +1588,35 @@ i32 search(
     }
 
     // LATE MOVE PRUNING
-    if (G(229, !in_check) && G(229, G(230, alpha) == G(230, beta - 1)) &&
-        G(229, quiets_evaluated > (G(231, 1) + G(231, depth * depth)) >>
+    if (G(233, !in_check) && G(233, G(234, alpha) == G(234, beta - 1)) &&
+        G(233, quiets_evaluated > (G(235, 1) + G(235, depth * depth)) >>
                    !improving)) {
       break;
     }
   }
 
   // MATE / STALEMATE DETECTION
-  if (G(232, best_score) == G(232, -inf)) {
-    return G(233, (ply - mate)) * G(233, in_check);
+  if (G(236, best_score) == G(236, -inf)) {
+    return G(237, (ply - mate)) * G(237, in_check);
   }
+
+  G(
+      238, // UPDATE CORRECTION HISTORY
+      if (G(239,
+            G(240, tt_flag) != G(240, (best_score < stack[ply].static_eval))) &&
+          G(239, G(241, stack[ply].best_move.takes_piece) == G(241, None))) {
+        G(242, i32 dd = depth * depth; if (dd > 64) { dd = 64; })
+        G(242, i32 target = best_score - stack[ply].static_eval; G(
+              243, if (target < -96) { target = -96; })
+              G(243, if (target > 96) { target = 96; }))
+
+        for (i32 i = 0; i < 2; i++) {
+          *corr_entries[i] =
+              (G(244, G(245, *corr_entries[i]) * G(245, (600 - dd))) +
+               G(244, G(246, target) * G(246, 256) * G(246, dd))) /
+              600;
+        }
+      })
 
   G(238, // UPDATE TRANSPOSITION TABLE
         *tt_entry = (TTEntry){.partial_hash = tt_hash_partial,
@@ -1598,20 +1624,6 @@ i32 search(
                               .score = best_score,
                               .depth = depth,
                               .flag = tt_flag};)
-
-  G(
-      238, // UPDATE CORRECTION HISTORY
-      if (G(236,
-            G(234, tt_flag) != G(234, (best_score < stack[ply].static_eval))) &&
-          G(236, G(235, stack[ply].best_move.takes_piece) == G(235, None))) {
-        G(237, i32 dd = depth * depth; if (dd > 64) { dd = 64; })
-        G(237, i32 target = best_score - stack[ply].static_eval; G(
-              239, if (target < -96) { target = -96; })
-              G(239, if (target > 96) { target = 96; }))
-
-        *material_entry =
-            (*material_entry * (600 - dd) + target * 256 * dd) / 600;
-      })
 
   return best_score;
 }
@@ -1632,13 +1644,13 @@ S(1) void init() {
       80, // INIT DIAGONAL MASKS
       for (i32 sq = 0; sq < 64; sq++) {
         const u64 bb = 1ULL << sq;
-        G(240, u64 sw_bb = southwest(bb);)
-        G(240, u64 ne_bb = northeast(bb);)
+        G(247, u64 sw_bb = southwest(bb);)
+        G(247, u64 ne_bb = northeast(bb);)
         for (i32 i = 6; i > 0; i--) {
-          G(241, sw_bb |= southwest(sw_bb);)
-          G(241, ne_bb |= northeast(ne_bb);)
+          G(248, sw_bb |= southwest(sw_bb);)
+          G(248, ne_bb |= northeast(ne_bb);)
         }
-        diag_mask[sq] = G(242, sw_bb) | G(242, ne_bb);
+        diag_mask[sq] = G(249, sw_bb) | G(249, ne_bb);
       })
 }
 
@@ -1695,7 +1707,7 @@ static void print_info(const Position *pos, const i32 depth, const i32 alpha,
   if (score > alpha && score < beta) {
     putl(" pv ");
     char move_name[8];
-    move_str(H(54, 2, &pv_move), H(54, 2, pos->flipped), H(54, 2, move_name));
+    move_str(H(50, 2, &pv_move), H(50, 2, pos->flipped), H(50, 2, move_name));
     putl(move_name);
   }
 
@@ -1716,18 +1728,18 @@ void iteratively_deepen(
   for (i32 depth = 1; depth < max_ply; depth++) {
 #endif
     // ASPIRATION WINDOWS
-    G(243, i32 window = 16;)
-    G(243, size_t elapsed;)
+    G(250, i32 window = 16;)
+    G(250, size_t elapsed;)
     while (true) {
-      G(244, const i32 beta = G(245, score) + G(245, window);)
-      G(244, const i32 alpha = score - window;)
+      G(251, const i32 beta = G(252, score) + G(252, window);)
+      G(251, const i32 alpha = score - window;)
       score = search(
 #ifdef FULL
           &data->nodes,
 #endif
-          H(178, 4, H(179, 4, &data->pos), H(179, 4, beta), H(179, 4, depth),
-            H(179, 4, data)),
-          H(178, 4, H(180, 4, 0), H(180, 4, false), H(180, 4, alpha)));
+          H(182, 4, H(183, 4, &data->pos), H(183, 4, beta), H(183, 4, depth),
+            H(183, 4, data)),
+          H(182, 4, H(184, 4, 0), H(184, 4, false), H(184, 4, alpha)));
 #ifdef FULL
       if (data->thread_id == 0) {
         print_info(&data->pos, depth, alpha, beta, score, data->nodes,
@@ -1735,10 +1747,10 @@ void iteratively_deepen(
       }
 #endif
       elapsed = get_time() - start_time;
-      G(246, window *= 2;)
+      G(253, window *= 2;)
       G(
-          246, if (G(247, elapsed > data->max_time) ||
-                   G(247, (G(248, score > alpha) && G(248, score < beta)))) {
+          253, if (G(254, elapsed > data->max_time) ||
+                   G(254, (G(255, score > alpha) && G(255, score < beta)))) {
             break;
           })
     }
@@ -1818,8 +1830,8 @@ void run_smp() {
   for (i32 i = 1; i < thread_count; i++) {
     ThreadHead *helper_head =
         (ThreadHead *)&thread_stacks[i][thread_stack_size - sizeof(ThreadHead)];
-    G(249, helper_head->data.pos = main_data->pos;)
-    G(249, helper_head->data.max_time = -1LL;)
+    G(256, helper_head->data.pos = main_data->pos;)
+    G(256, helper_head->data.max_time = -1LL;)
 #ifdef FULL
     helper_head->data.thread_id = i;
 #endif
@@ -1840,8 +1852,8 @@ void run_smp() {
 #endif
 
   char move_name[8];
-  move_str(H(54, 3, &main_data->stack[0].best_move),
-           H(54, 3, main_data->pos.flipped), H(54, 3, move_name));
+  move_str(H(50, 3, &main_data->stack[0].best_move),
+           H(50, 3, main_data->pos.flipped), H(50, 3, move_name));
   putl("bestmove ");
   puts(move_name);
 }
@@ -1856,7 +1868,7 @@ S(1) void display_pos(Position *const pos) {
     for (i32 file = 0; file < 8; file++) {
       i32 sq = rank * 8 + file;
       u64 bb = 1ULL << sq;
-      i32 piece = piece_on(H(52, 9, &npos), H(52, 9, sq));
+      i32 piece = piece_on(H(55, 9, &npos), H(55, 9, sq));
       if (bb & npos.colour[0]) {
         if (piece == Pawn) {
           putl("P");
@@ -1954,8 +1966,8 @@ S(1) void run() {
   setvbuf(stdout, NULL, _IONBF, 0);
 #endif
 
-  G(250, char line[4096];)
-  G(250, init();)
+  G(257, char line[4096];)
+  G(257, init();)
 #ifdef FULL
   main_data = calloc(1, sizeof(ThreadData));
   tt = malloc(tt_length * sizeof(TTEntry));
@@ -2030,8 +2042,8 @@ S(1) void run() {
              elapsed, nps);
     }
 #endif
-    G(251, if (G(257, line[0]) == G(257, 'q')) { exit_now(); })
-    else G(251, if (G(252, line[0]) == G(252, 'g')) {
+    G(258, if (G(261, line[0]) == G(261, 'i')) { puts("readyok"); })
+    else G(258, if (G(260, line[0]) == G(260, 'g')) {
       stop = false;
 #ifdef FULL
       while (true) {
@@ -2062,9 +2074,8 @@ S(1) void run() {
       run_smp();
 #endif
     })
-    else G(251, if (G(253, line[0]) == G(253, 'i')) { puts("readyok"); })
-    else G(251, if (G(254, line[0]) == G(254, 'p')) {
-      G(255, main_data->pos = start_pos;)
+    else G(258, if (G(262, line[0]) == G(262, 'p')) {
+      G(263, main_data->pos = start_pos;)
         while (true) {
           bool line_continue = getl(line);
 
@@ -2081,11 +2092,11 @@ S(1) void run() {
               H(95, 4, false), H(95, 4, moves));
             for (i32 i = 0; i < num_moves; i++) {
               char move_name[8];
-              move_str(H(54, 4, &moves[i]), H(54, 4, main_data->pos.flipped),
-                H(54, 4, move_name));
+              move_str(H(50, 4, &moves[i]), H(50, 4, main_data->pos.flipped),
+                H(50, 4, move_name));
               assert(move_string_equal(line, move_name) ==
                 !strcmp(line, move_name));
-              if (move_string_equal(G(256, move_name), G(256, line))) {
+              if (move_string_equal(G(264, move_name), G(264, line))) {
                 makemove(H(80, 4, &main_data->pos), H(80, 4, &moves[i]));
                 break;
               }
@@ -2096,6 +2107,7 @@ S(1) void run() {
           }
         }
     })
+    else G(258, if (G(259, line[0]) == G(259, 'q')) { exit_now(); })
   }
 }